requirements.txt
opencv-python
numpy
pytesseract
pyserial
PyYAML